Criando contexto global para  compartilhar props entre páginas diferentes, o que é diferente de compartilhar entre componentes.
