+++ aulas/reactjs/aulas/02-Criando-SPAS/02-ignite-timer/src/App.tsx
@@ -4,13 +4,16 @@ import { defaultTheme } from './styles/themes/default'
 import { GlobalStyle } from './styles/global'
 import { Router } from './Router'
 import { BrowserRouter } from 'react-router-dom'
+import { CyclesContextProvider } from './contexts/CyclesContext'
 
 export function App() {
   return (
     <>
       <ThemeProvider theme={defaultTheme}>
         <BrowserRouter>
-          <Router />
+          <CyclesContextProvider>
+            <Router />
+          </CyclesContextProvider>
           <GlobalStyle />
         </BrowserRouter>
       </ThemeProvider>
