+++ aulas/reactjs/aulas/02-Criando-SPAS/02-ignite-timer/src/contexts/CyclesContext.tsx
@@ -1,0 +1,103 @@
+import { createContext, ReactNode, useState } from 'react'
+
+interface Cycle {
+  id: string
+  task: string
+  minutesAmount: number
+  startDate: Date
+  interruptedDate?: Date
+  finishedDate?: Date
+}
+
+interface CreateCycleData {
+  task: string
+  minutesAmount: number
+}
+
+interface CyclesContextType {
+  cycles: Cycle[]
+  activeCycle: Cycle | undefined
+  activeCycleId: string | null
+  markCurrentCycleAsFinished: () => void
+  amountSecondsPassed: number
+  setSecondsPassed: (seconds: number) => void
+  createNewCycle: (data: CreateCycleData) => void
+  interrupCurrentCycle: () => void
+}
+
+interface CyclesContextProviderProps {
+  children: ReactNode
+}
+
+export const CyclesContext = createContext({} as CyclesContextType)
+
+export function CyclesContextProvider({
+  children,
+}: CyclesContextProviderProps) {
+  const [cycles, setCycles] = useState<Cycle[]>([])
+  const [activeCycleId, setActiveCycleId] = useState<string | null>(null)
+  const [amountSecondsPassed, setamountSecondsPassed] = useState(0)
+
+  const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId)
+
+  function createNewCycle(data: CreateCycleData) {
+    const newCycle: Cycle = {
+      id: String(new Date().getTime()),
+      task: data.task,
+      minutesAmount: data.minutesAmount,
+      startDate: new Date(),
+    }
+
+    setCycles((state) => [...state, newCycle]) // closures, faço no formato de function
+    setActiveCycleId(newCycle.id)
+    setamountSecondsPassed(0)
+    // reset()
+  }
+
+  function interrupCurrentCycle() {
+    setCycles((state) =>
+      state.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return { ...cycle, interruptedDate: new Date() }
+        } else {
+          return cycle
+        }
+      }),
+    )
+
+    setActiveCycleId(null)
+  }
+
+  function markCurrentCycleAsFinished() {
+    setCycles((state) =>
+      state.map((cycle) => {
+        if (cycle.id === activeCycleId) {
+          return { ...cycle, finishedDate: new Date() }
+        } else {
+          return cycle
+        }
+      }),
+    )
+  }
+
+  function setSecondsPassed(seconds: number) {
+    setamountSecondsPassed(seconds)
+  }
+
+  return (
+    <CyclesContext.Provider
+      value={{
+        cycles,
+        activeCycle,
+        activeCycleId,
+        markCurrentCycleAsFinished,
+        amountSecondsPassed,
+        setSecondsPassed,
+        createNewCycle,
+        interrupCurrentCycle,
+      }}
+    >
+      {children}
+    </CyclesContext.Provider>
+  )
+}
